refactor!: Turn NonBlockingSocket into trait to allow alternate socket types
The session are now generic over the underlying socket type and is
passed an existing socket instead of creating one itself.

Some types had to be made public to make this work, for instance
UdpMessage is now a public (opaque-ish) type, implementors of the socket
trait should just serialize/deserialize it and pass it on to the
underlying socket.